+++ pleo-antaeus-core/src/test/kotlin/io/pleo/antaeus/core/services/BillingServiceTest.kt
@@ -7,10 +7,8 @@ import io.pleo.antaeus.core.exceptions.CurrencyMismatchException
 import io.pleo.antaeus.core.exceptions.CustomerNotFoundException
 import io.pleo.antaeus.core.exceptions.NetworkException
 import io.pleo.antaeus.core.external.PaymentProvider
-import io.pleo.antaeus.models.Currency
-import io.pleo.antaeus.models.Invoice
-import io.pleo.antaeus.models.InvoiceStatus
-import io.pleo.antaeus.models.Money
+import io.pleo.antaeus.data.constants.DEFAULT_PAGE_SIZE
+import io.pleo.antaeus.models.*
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import java.math.BigDecimal
@@ -22,12 +20,28 @@ class BillingServiceTest {
 
     @Test
     fun `should call paymentProvider with list of pending invoices and update all of them to PAID`() {
-        val invoices = expectListOfInvoicesSuitableForBeingPaid("PENDING")
+        val invoicesPage = expectListOfInvoicesSuitableForBeingPaid("PENDING")
         billingService.processInvoicesByStatus("PENDING")
+        for (invoice in invoicesPage.invoices) {
+            verify {
+                paymentProvider.charge(invoice)
+                invoiceService.updateStatus(invoice.id, InvoiceStatus.PAID.toString())
+                invoiceService.fetchPageByStatus(status = "PENDING", pageSize = any(), marker = any())
+            }
+        }
+    }
+
+    @Test
+    fun `should get the invoice list using pagination and call paymentProvider`() {
+        val invoices = expectInvoicePageof1ElementToBePaid("PENDING")
+        billingService.processInvoicesByStatus("PENDING", pageSize = 1)
         for (invoice in invoices) {
             verify {
                 paymentProvider.charge(invoice)
                 invoiceService.updateStatus(invoice.id, InvoiceStatus.PAID.toString())
+            }
+            verify(exactly = 2) {
+                invoiceService.fetchPageByStatus(status = "PENDING", pageSize = 1, marker = any())
             }
         }
     }
@@ -127,15 +141,15 @@ class BillingServiceTest {
     }
 
 
-    private fun expectListOfInvoicesSuitableForBeingPaid(status: String): List<Invoice> {
-        val expectedInvoices = aListOfInvoices(status = status)
+    private fun expectListOfInvoicesSuitableForBeingPaid(status: String): InvoicePage {
+        val expectedInvoices = aPageOfInvoices(aListOfInvoices(status = status), true)
         every {
-            invoiceService.fetchByStatus(status)
+            invoiceService.fetchPageByStatus(status, pageSize = DEFAULT_PAGE_SIZE, marker = null)
         } returns expectedInvoices
         every {
             paymentProvider.charge(any())
         } returns true
-        for(invoice in expectedInvoices) {
+        for(invoice in expectedInvoices.invoices) {
             every {
                 paymentProvider.charge(invoice)
             } returns true
@@ -144,6 +158,36 @@ class BillingServiceTest {
             } returns invoice
         }
         return expectedInvoices
+    }
+
+    private fun expectInvoicePageof1ElementToBePaid(status: String): List<Invoice> {
+        val invoicesList = aListOfInvoices(size = 2, status = status)
+        every {
+            invoiceService.fetchPageByStatus(status, pageSize = 1, marker = null)
+        } returns InvoicePage(
+            invoices = listOf(invoicesList.first()),
+            isLast = false,
+            marker = invoicesList.first().id
+        )
+        every {
+            invoiceService.fetchPageByStatus(status, pageSize = 1, marker = invoicesList.first().id)
+        } returns InvoicePage(
+            invoices = listOf(invoicesList.last()),
+            isLast = true,
+            marker = invoicesList.last().id
+        )
+        every {
+            paymentProvider.charge(any())
+        } returns true
+        for(invoice in invoicesList) {
+            every {
+                paymentProvider.charge(invoice)
+            } returns true
+            every {
+                invoiceService.updateStatus(invoice.id, InvoiceStatus.PAID.toString())
+            } returns invoice
+        }
+        return invoicesList
     }
 
     private fun expectAlreadyPaidInvoice(id: Int): Invoice {
@@ -160,8 +204,8 @@ class BillingServiceTest {
         } throws CustomerNotFoundException(invoice.customerId)
 
         every {
-            invoiceService.fetchByStatus(InvoiceStatus.PENDING.toString())
-        } returns listOf(invoice)
+            invoiceService.fetchPageByStatus(InvoiceStatus.PENDING.toString(), pageSize = DEFAULT_PAGE_SIZE, marker = null)
+        } returns InvoicePage(listOf(invoice), isLast = true, marker = invoice.id)
         every {
             invoiceService.fetch(invoice.id)
         } returns invoice
@@ -176,8 +220,8 @@ class BillingServiceTest {
         } throws CurrencyMismatchException(invoice.customerId, invoice.customerId)
 
         every {
-            invoiceService.fetchByStatus(InvoiceStatus.PENDING.toString())
-        } returns listOf(invoice)
+            invoiceService.fetchPageByStatus(InvoiceStatus.PENDING.toString(), pageSize = DEFAULT_PAGE_SIZE, marker = null)
+        } returns InvoicePage(listOf(invoice), isLast = true, marker = invoice.id)
         every {
             invoiceService.fetch(invoice.id)
         } returns invoice
@@ -212,6 +256,9 @@ class BillingServiceTest {
             invoiceService.updateStatus(invoice.id, InvoiceStatus.PENDING.toString())
         } returns invoice
     }
+
+    private fun aPageOfInvoices(invoices: List<Invoice>, isLast: Boolean = true): InvoicePage =
+        InvoicePage(invoices, isLast, invoices.last().id)
 
     private fun aListOfInvoices(size: Int = 2, status: String = InvoiceStatus.PENDING.toString()): List<Invoice> =
             List(size) { anInvoice(it, InvoiceStatus.valueOf(status)) }