+++ pleo-antaeus-data/src/main/kotlin/io/pleo/antaeus/data/AntaeusDal.kt
@@ -7,94 +7,24 @@
 
 package io.pleo.antaeus.data
 
-import io.pleo.antaeus.models.Currency
-import io.pleo.antaeus.models.Customer
-import io.pleo.antaeus.models.Invoice
-import io.pleo.antaeus.models.InvoiceStatus
-import io.pleo.antaeus.models.Money
+import io.pleo.antaeus.models.*
 import org.jetbrains.exposed.sql.*
-import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
-import org.jetbrains.exposed.sql.transactions.transaction
+import io.pleo.antaeus.data.constants.DEFAULT_PAGE_SIZE
 
-class AntaeusDal(private val db: Database) {
-    fun fetchInvoice(id: Int): Invoice? {
-        // transaction(db) runs the internal query as a new database transaction.
-        return transaction(db) {
-            // Returns the first invoice with matching id.
-            InvoiceTable
-                .select { InvoiceTable.id.eq(id) }
-                .firstOrNull()
-                ?.toInvoice()
-        }
-    }
+class AntaeusDal(db: Database) {
+    private val invoiceDal = InvoiceDal(db)
+    private val customerDal = CustomerDal(db)
 
-    fun fetchInvoices(): List<Invoice> {
-        return transaction(db) {
-            InvoiceTable
-                .selectAll()
-                .map { it.toInvoice() }
-        }
-    }
+    fun fetchInvoice(id: Int): Invoice? = invoiceDal.fetchInvoice(id)
+    fun fetchInvoices(): List<Invoice> = invoiceDal.fetchInvoices()
+    fun fetchInvoicesByStatus(status: InvoiceStatus): List<Invoice> = invoiceDal.fetchInvoicesByStatus(status)
+    fun createInvoice(amount: Money, customer: Customer, status: InvoiceStatus = InvoiceStatus.PENDING): Invoice? =
+        invoiceDal.createInvoice(amount, customer, status)
+    fun updateInvoiceStatus(id: Int, newStatus: InvoiceStatus): Invoice? = invoiceDal.updateInvoiceStatus(id, newStatus)
+    fun fetchInvoicePagesByStatus(status: InvoiceStatus, pageSize: Int?, marker: Int?): InvoicePage =
+        invoiceDal.fetchInvoicesPageByStatus(status, pageSize.let { DEFAULT_PAGE_SIZE }, marker)
 
-    fun fetchInvoicesByStatus(status: InvoiceStatus): List<Invoice> {
-        return transaction (db) {
-            InvoiceTable
-                    .select(InvoiceTable.status.eq(status.toString()))
-                    .map { it.toInvoice() }
-        }
-    }
-
-    fun createInvoice(amount: Money, customer: Customer, status: InvoiceStatus = InvoiceStatus.PENDING): Invoice? {
-        val id = transaction(db) {
-            // Insert the invoice and returns its new id.
-            InvoiceTable
-                .insert {
-                    it[this.value] = amount.value
-                    it[this.currency] = amount.currency.toString()
-                    it[this.status] = status.toString()
-                    it[this.customerId] = customer.id
-                } get InvoiceTable.id
-        }
-
-        return fetchInvoice(id)
-    }
-
-    fun updateInvoiceStatus(id: Int, newStatus: InvoiceStatus): Invoice? {
-        val id = transaction(db) {
-            InvoiceTable
-                    .update( { InvoiceTable.id.eq(id) } ) {
-                        it[status] = newStatus.toString()
-                    }
-
-        }
-        return fetchInvoice(id)
-    }
-
-    fun fetchCustomer(id: Int): Customer? {
-        return transaction(db) {
-            CustomerTable
-                .select { CustomerTable.id.eq(id) }
-                .firstOrNull()
-                ?.toCustomer()
-        }
-    }
-
-    fun fetchCustomers(): List<Customer> {
-        return transaction(db) {
-            CustomerTable
-                .selectAll()
-                .map { it.toCustomer() }
-        }
-    }
-
-    fun createCustomer(currency: Currency): Customer? {
-        val id = transaction(db) {
-            // Insert the customer and return its new id.
-            CustomerTable.insert {
-                it[this.currency] = currency.toString()
-            } get CustomerTable.id
-        }
-
-        return fetchCustomer(id)
-    }
+    fun fetchCustomer(id: Int): Customer? = customerDal.fetchCustomer(id)
+    fun fetchCustomers(): List<Customer> = customerDal.fetchCustomers()
+    fun createCustomer(currency: Currency): Customer? = customerDal.createCustomer(currency)
 }