+++ pleo-antaeus-data/src/main/kotlin/io/pleo/antaeus/data/Constants.kt
@@ -1,0 +1,3 @@
+package io.pleo.antaeus.data.constants
+
+const val DEFAULT_PAGE_SIZE = 100