+++ pleo-antaeus-models/src/main/kotlin/io/pleo/antaeus/models/InvoicePage.kt
@@ -1,0 +1,7 @@
+package io.pleo.antaeus.models
+
+data class InvoicePage (
+    val invoices: List<Invoice>,
+    val isLast: Boolean,
+    val marker: Int?
+)