+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
-# 1.0.0 Initial Version
+## 1.1.0
+* Included circuit breaker support (https://resilience4j.readme.io/docs/getting-started-4) for calling payment provider
+* Included keyset pagination (https://use-the-index-luke.com/no-offset) for iterating over invoices list
+
+## 1.0.0 Initial Version
 * Initial version using coroutines to process the list of invoices
 * Using krontab (https://github.com/InsanusMokrassar/krontab) to schedule the job